+++ apina-core/src/main/resources/typescript/runtime-angular2.ts
@@ -1,0 +1,177 @@
+export interface RequestData {
+    uriTemplate: string;
+    method: string;
+    pathVariables?: any;
+    requestParams?: any;
+    requestBody?: any;
+    responseType?: string;
+}
+
+export interface Serializer {
+    serialize(o: any): any;
+    deserialize(o: any): any;
+}
+
+const identitySerializer: Serializer = {
+    serialize(o) {
+        return o;
+    },
+    deserialize(o) {
+        return o;
+    }
+};
+
+function enumSerializer(enumObject: any): Serializer {
+    return {
+        serialize(o) {
+            if (o === null || o === undefined)
+                return o;
+            else
+                return enumObject[o];
+        },
+        deserialize(o) {
+            if (o === null || o === undefined)
+                return o;
+            else
+                return enumObject[o];
+        }
+    }
+}
+
+interface SerializerMap {
+    [name: string]: Serializer;
+}
+
+@Injectable()
+export class ApinaConfig {
+
+    /** Prefix added for all API calls */
+    baseUrl: string = "";
+
+    private serializers: SerializerMap = {
+        any: identitySerializer,
+        string: identitySerializer,
+        number: identitySerializer,
+        boolean: identitySerializer
+    };
+
+    constructor() {
+        registerDefaultSerializers(this);
+    }
+
+    serialize(value: any, type: string): any {
+        return this.lookupSerializer(type).serialize(value);
+    }
+
+    deserialize(value: any, type: string): any {
+        return this.lookupSerializer(type).deserialize(value);
+    }
+
+    registerSerializer(name: string, serializer: Serializer) {
+        this.serializers[name] = serializer;
+    }
+
+    registerEnumSerializer(name: string, enumObject: any) {
+        this.registerSerializer(name, enumSerializer(enumObject));
+    }
+
+    registerClassSerializer(name: string, fields: any) {
+        this.registerSerializer(name, this.classSerializer(fields));
+    }
+
+    registerIdentitySerializer(name: string) {
+        this.registerSerializer(name, identitySerializer);
+    }
+
+    private classSerializer(fields: any): Serializer {
+        function mapProperties(obj: any, propertyMapper: (value: any, type: string) => any) {
+            if (obj === null || obj === undefined) {
+                return obj;
+            }
+
+            const result: any = {};
+
+            for (const name in fields) {
+                if (fields.hasOwnProperty(name)) {
+                    const value: any = obj[name];
+                    const type: string = fields[name];
+                    result[name] = propertyMapper(value, type);
+                }
+            }
+
+            return result;
+        }
+
+        const serialize = this.serialize.bind(this);
+        const deserialize = this.deserialize.bind(this);
+        return {
+            serialize(obj) {
+                return mapProperties(obj, serialize);
+            },
+            deserialize(obj) {
+                return mapProperties(obj, deserialize);
+            }
+        };
+    }
+
+    private lookupSerializer(type: string): Serializer {
+        if (!type) throw new Error("no type given");
+
+        if (type.indexOf('[]', type.length - 2) !== -1) { // type.endsWith('[]')
+            const elementType = type.substring(0, type.length - 2);
+            const elementSerializer = this.lookupSerializer(elementType);
+            return arraySerializer(elementSerializer);
+        }
+        const serializer = this.serializers[type];
+        if (serializer) {
+            return serializer;
+        } else {
+            throw new Error(`could not find serializer for type '${type}'`);
+        }
+    }
+}
+
+@Injectable()
+export class ApinaEndpointContext {
+
+    constructor(private http: Http, private config: ApinaConfig) {
+    }
+
+    request(data: RequestData): Observable<any> {
+        const url = this.buildUrl(data.uriTemplate, data.pathVariables);
+
+        return this.http.request(url, { method: data.method, search: data.requestParams, body: data.requestBody })
+            .map(r => data.responseType ? this.deserialize(r.json(), data.responseType) : r)
+            .share();
+    }
+
+    serialize(value: any, type: string): any {
+        return this.config.serialize(value, type);
+    }
+
+    deserialize(value: any, type: string): any {
+        return this.config.deserialize(value, type);
+    }
+
+    private buildUrl(uriTemplate: String, pathVariables: any): string {
+        return this.config.baseUrl + uriTemplate.replace(/\{([^}]+)}/g, (match, name) => pathVariables[name]);
+    }
+}
+
+function arraySerializer(elementSerializer: Serializer): Serializer {
+    function safeMap(value: any[], mapper: (a: any) => any) {
+        if (!value)
+            return value;
+        else
+            return value.map(mapper);
+    }
+
+    return {
+        serialize(value) {
+            return safeMap(value, elementSerializer.serialize.bind(elementSerializer));
+        },
+        deserialize(value) {
+            return safeMap(value, elementSerializer.deserialize.bind(elementSerializer));
+        }
+    }
+}